requirements.txt
# FINAL-MVP-CATTLE Requirements
# Tested with Python 3.11.x + CUDA 12.4

# PyTorch (installed separately via pip with CUDA - see setup.bat)
# torch>=2.2.0
# torchvision>=0.17.0

# ML & Computer Vision
ultralytics>=8.1.0
opencv-python>=4.9.0
timm>=0.9.12
scikit-learn>=1.4.0
albumentations>=1.3.1

# Data & Visualization
numpy>=1.26.0,<2.0
pandas>=2.1.0
matplotlib>=3.8.0
seaborn>=0.13.0
pillow>=10.2.0

# Utilities
tqdm>=4.66.0
tensorboard>=2.15.0

# Web App
streamlit>=1.30.0
